broadcast-common: improve output messages
The style with caps and period at the end is now consistent with the style
that git and yadm use in the case where everything is up to date.

Some outputs were factually wrong (mentioning network hops we no longer do).

The style of reporting assumptions is now more consistent.